July 15th 2020 2:03PM EST

+++ SDLMan.h
@@ -110,5 +110,5 @@ private:
 	Uint32 fpsCount;
 
 	// the value you want
-	float fps;
+	decimal fps;
 };
